Update ChatBot.java
Line 51, and 52 are all new lines of code which process the input of the user. However, we have to write the logic for keyword detection.  We are able to store the string input of the user. Next, we need to see if the user input contains a keyword from the keywords.txt. If it does then we have a scripted output for each keyword.